Split Vexer into SBCS and Literate
